+++ tests/models/test_hooks.py
@@ -274,8 +274,8 @@ class HookedModel(BoringModel):
     @staticmethod
     def _auto_train_batch(trainer, model, batches, device=torch.device("cpu"), current_epoch=0, **kwargs):
         using_native_amp = kwargs.get("amp_backend") == "native"
-        using_deepspeed = kwargs.get("plugins") == "deepspeed"
-        using_plugin = kwargs.get("amp_backend") or kwargs.get("plugins")
+        using_deepspeed = kwargs.get("strategy") == "deepspeed"
+        using_plugin = kwargs.get("amp_backend") or kwargs.get("strategy")
         out = []
         on_before_optimizer_step = [
             dict(name="Callback.on_before_optimizer_step", args=(trainer, model, ANY, 0)),
@@ -339,8 +339,8 @@ class HookedModel(BoringModel):
 
     @staticmethod
     def _manual_train_batch(trainer, model, batches, device=torch.device("cpu"), **kwargs):
-        using_deepspeed = kwargs.get("plugins") == "deepspeed"
-        using_plugin = kwargs.get("amp_backend") or kwargs.get("plugins")
+        using_deepspeed = kwargs.get("strategy") == "deepspeed"
+        using_plugin = kwargs.get("amp_backend") or kwargs.get("strategy")
         out = []
         for i in range(batches):
             out.extend(
@@ -438,14 +438,14 @@ class HookedModel(BoringModel):
 @RunIf(deepspeed=True, min_gpus=1, special=True)
 def test_trainer_model_hook_system_fit_deepspeed_automatic_optimization(tmpdir):
     _run_trainer_model_hook_system_fit(
-        dict(gpus=1, precision=16, plugins="deepspeed"), tmpdir, automatic_optimization=True
+        dict(gpus=1, precision=16, strategy="deepspeed"), tmpdir, automatic_optimization=True
     )
 
 
 @RunIf(deepspeed=True, min_gpus=1, special=True)
 def test_trainer_model_hook_system_fit_deepspeed_manual_optimization(tmpdir):
     _run_trainer_model_hook_system_fit(
-        dict(gpus=1, precision=16, plugins="deepspeed"), tmpdir, automatic_optimization=False
+        dict(gpus=1, precision=16, strategy="deepspeed"), tmpdir, automatic_optimization=False
     )
 
 
@@ -521,18 +521,18 @@ def _run_trainer_model_hook_system_fit(kwargs, tmpdir, automatic_optimization):
         dict(name="configure_callbacks"),
         dict(name="Callback.on_before_accelerator_backend_setup", args=(trainer, model)),
         # DeepSpeed needs the batch size to figure out throughput logging
-        *([dict(name="train_dataloader")] if kwargs.get("plugins") == "deepspeed" else []),
+        *([dict(name="train_dataloader")] if kwargs.get("strategy") == "deepspeed" else []),
         dict(name="Callback.setup", args=(trainer, model), kwargs=dict(stage="fit")),
         dict(name="setup", kwargs=dict(stage="fit")),
         dict(name="configure_sharded_model"),
         dict(name="Callback.on_configure_sharded_model", args=(trainer, model)),
         # DeepSpeed skips initializing optimizers here as they are handled via config
-        *([dict(name="configure_optimizers")] if kwargs.get("plugins") != "deepspeed" else []),
+        *([dict(name="configure_optimizers")] if kwargs.get("strategy") != "deepspeed" else []),
         dict(name="Callback.on_fit_start", args=(trainer, model)),
         dict(name="on_fit_start"),
         # TODO: explore whether DeepSpeed can have the same flow for optimizers
         # DeepSpeed did not find any optimizer in the config so they are loaded here
-        *([dict(name="configure_optimizers")] if kwargs.get("plugins") == "deepspeed" else []),
+        *([dict(name="configure_optimizers")] if kwargs.get("strategy") == "deepspeed" else []),
         dict(name="Callback.on_pretrain_routine_start", args=(trainer, model)),
         dict(name="on_pretrain_routine_start"),
         dict(name="Callback.on_pretrain_routine_end", args=(trainer, model)),